// app/src/app/core/models/erro.model.ts
export class Erro {
  errors: any;
  message: string;
  result: any;
  url: any;
  status: any;
  statusCode: any;

// constructor(message:string,errors:any,statusCode:string,url:string,data:any){
//   this.message = message;
//   this.errors = errors;
//   this.statusCode = statusCode;
//   this.url = url;
//   this.data = data;
// }
}